fastapi-project/pyproject.toml
[tool.poetry]
name = "fastapi-project"
version = "0.1.0"
description = ""
authors = ["hotratx <mgviero@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_project"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
alembic = "^1.9.1"
python-dotenv = "^0.20.0"
sqlmodel = "^0.0.8"
asyncpg = "^0.27.0"
pydantic = {extras = ["email"], version = "^1.10.2"}
aioredis = "^2.0.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-asyncio = "^0.20.3"
httpx = "^0.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
